requirements.txt
streamlit
openai
python-dotenv
pyyaml
streamlit-feedback